+++ ExotischNederland/ExotischNederland.csproj
@@ -49,11 +49,11 @@
     <Compile Include="Menus\AreaMenu.cs" />
     <Compile Include="Menus\GameMenu.cs" />
     <Compile Include="Menus\IMenu.cs" />
-    <Compile Include="Menus\MainMenu.cs" />
+    <Compile Include="Menus\AuthenticationMenu.cs" />
     <Compile Include="Menus\ObservationMenu.cs" />
-    <Compile Include="Menus\UserManagementMenu.cs" />
-    <Compile Include="Menus\RouteMenu.cs" />
     <Compile Include="Menus\UserMenu.cs" />
+    <Compile Include="Menus\RouteMenu.cs" />
+    <Compile Include="Menus\MainMenu.cs" />
     <Compile Include="Models\Answer.cs" />
     <Compile Include="Models\Area.cs" />
     <Compile Include="Models\Game.cs" />
